+++ adminSiteServer/serverUtil.tsx
@@ -3,39 +3,7 @@ import * as lodash from "lodash"
 import { quote } from "shell-quote"
 import urljoin from "url-join"
 import { ADMIN_BASE_URL } from "settings"
-import * as shell from "shelljs"
 import * as util from "util"
-
-interface ExecReturn {
-    code: number
-    stdout: string
-    stderr: string
-}
-
-export class ExecError extends Error implements ExecReturn {
-    code: number
-    stdout: string
-    stderr: string
-
-    constructor(props: ExecReturn) {
-        super(props.stderr)
-        this.code = props.code
-        this.stdout = props.stdout
-        this.stderr = props.stderr
-    }
-}
-
-export const exec = (
-    command: string,
-    options?: shell.ExecOptions
-): Promise<ExecReturn> =>
-    new Promise((resolve, reject) => {
-        shell.exec(command, options || {}, (code, stdout, stderr) =>
-            code === 0
-                ? resolve({ code, stdout, stderr })
-                : reject(new ExecError({ code, stdout, stderr }))
-        )
-    })
 
 // Exception format that can be easily given as an API error
 export class JsonError extends Error {
@@ -87,5 +55,5 @@ export const execFormatted = async (
 ) => {
     const formatCmd = util.format(cmd, ...args.map((s) => quote([s])))
     if (verbose) console.log(formatCmd)
-    return await exec(formatCmd, { silent: !verbose })
+    return await execWrapper(formatCmd, { silent: !verbose })
 }
