+++ db/exportMetadata.ts
@@ -5,7 +5,7 @@ import * as fs from "fs-extra"
 import parseArgs from "minimist"
 
 import { DB_NAME, DB_USER, DB_PASS, DB_HOST, DB_PORT } from "serverSettings"
-import { exec } from "adminSiteServer/serverUtil"
+import { execWrapper } from "./execWrapper"
 
 const argv = parseArgs(process.argv.slice(2))
 const withPasswords = argv["with-passwords"]
@@ -33,12 +33,12 @@ async function dataExport() {
     process.env.MYSQL_PWD = DB_PASS
 
     // Dump all tables including schema but exclude the rows of data_values
-    await exec(
+    await execWrapper(
         `mysqldump --default-character-set=utf8mb4 --no-tablespaces -u '${DB_USER}' -h '${DB_HOST}' -P ${DB_PORT} ${DB_NAME} ${excludeTables
             .map((tableName) => `--ignore-table=${DB_NAME}.${tableName}`)
             .join(" ")} -r ${filePath}`
     )
-    await exec(
+    await execWrapper(
         `mysqldump --default-character-set=utf8mb4 --no-tablespaces -u '${DB_USER}' -h '${DB_HOST}' -P ${DB_PORT} --no-data ${DB_NAME} ${excludeTables.join(
             " "
         )} >> ${filePath}`
@@ -46,7 +46,7 @@ async function dataExport() {
 
     if (!withPasswords) {
         // Strip passwords
-        await exec(`sed -i -e "s/bcrypt[^']*//g" ${filePath}`)
+        await execWrapper(`sed -i -e "s/bcrypt[^']*//g" ${filePath}`)
         // Add default admin user
         await fs.appendFile(
             filePath,
