+++ db/tsconfig.json
@@ -3,6 +3,5 @@
     "compilerOptions": {
         "outDir": "./compiled/"
     },
-    "exclude": ["*"],
     "references": [{ "path": "../clientUtils" }]
 }
